Move inspect command into own file
